move AddrList to pstoremen, unexport it
